move Settings / DeviceSetting / DeviceSettings to types/device-settings.mts
Those three types describe per-device Homey settings — the values that
ship in the PUT /settings/devices body and live in each device's own
`getSettings()` shape. They are not app-level settings. Keep
`types/app-settings.mts` for `HomeySettings` only, which is the
app-scoped Homey Settings store (context key, tokens, prefixed
username/password).
