+++ types/app-settings.mts
@@ -1,9 +1,3 @@
-import type { BaseSettings } from './bases.mts'
-
-export type DeviceSetting = Record<string, Settings[keyof Settings]>
-
-export type DeviceSettings = Record<string, DeviceSetting>
-
 export interface HomeySettings {
   readonly contextKey?: string | null
   readonly expiry?: string | null
@@ -15,8 +9,4 @@ export interface HomeySettings {
   readonly notifiedVersion?: string | null
   readonly password?: string | null
   readonly username?: string | null
-}
-
-export interface Settings extends BaseSettings {
-  readonly always_on?: boolean
 }
